+++ src/test/java/com/codeit/monew/domain/article/service/ArticleServiceTest.java
@@ -8,7 +8,7 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 
-import com.codeit.monew.domain.article.dto.ArticleDto;
+import com.codeit.monew.domain.article.dto.response.ArticleDto;
 import com.codeit.monew.domain.article.entity.Article;
 import com.codeit.monew.domain.article.ArticleSource;
 import com.codeit.monew.domain.article.mapper.ArticleMapper;
